+++ src/App.tsx
@@ -1,56 +1,17 @@
 import React from 'react';
+import { Counter } from './components/Counter';
 
 type State = {
   count: number,
 };
 
+// eslint-disable-next-line react/prefer-stateless-function
 class App extends React.Component<{}, State> {
-  state = {
-    count: 0,
-  };
-
-  incrementCount = (num: number) => {
-    this.setState((prevState) => ({ count: prevState.count + num }));
-  };
-
-  addOne = () => {
-    this.incrementCount(1);
-  };
-
-  add100 = () => {
-    this.incrementCount(100);
-  };
-
-  increase = () => {
-    if (this.state.count % 5 === 0) {
-      this.incrementCount(100);
-    }
-
-    this.incrementCount(1);
-  };
-
-  render() {
-    const { count } = this.state;
-
-    const btnIncrement1 = React.createElement('button',
-      { type: 'button', onClick: this.addOne }, 'addOne');
-
-    const btnIncrement2 = React.createElement('button',
-      { type: 'button', onClick: this.add100 }, 'add100');
-
-    const btnIncrement3 = React.createElement('button',
-      { type: 'button', onClick: this.increase }, 'increase');
-
+  render(): React.ReactNode {
     return (
-      <>
-        <h1>
-          Count:
-          {count}
-        </h1>
-        {btnIncrement1}
-        {btnIncrement2}
-        {btnIncrement3}
-      </>
+      <div className="app">
+        <Counter />
+      </div>
     );
   }
 }
